relocate sf, dialog boxes css
