+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\ayubi\\SafeRoute\\src\\index.js":"1","C:\\Users\\ayubi\\SafeRoute\\src\\reportWebVitals.js":"2","C:\\Users\\ayubi\\SafeRoute\\src\\App.js":"3","C:\\Users\\ayubi\\SafeRoute\\src\\components\\Navbar\\index.js":"4","C:\\Users\\ayubi\\SafeRoute\\src\\components\\SearchField\\index.js":"5"},{"size":500,"mtime":1608424109636,"results":"6","hashOfConfig":"7"},{"size":362,"mtime":1608424109636,"results":"8","hashOfConfig":"7"},{"size":1329,"mtime":1609211383169,"results":"9","hashOfConfig":"7"},{"size":6275,"mtime":1609308660287,"results":"10","hashOfConfig":"7"},{"size":788,"mtime":1609211344172,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"1k0mzzr",{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"C:\\Users\\ayubi\\SafeRoute\\src\\index.js",[],["23","24"],"C:\\Users\\ayubi\\SafeRoute\\src\\reportWebVitals.js",[],"C:\\Users\\ayubi\\SafeRoute\\src\\App.js",[],"C:\\Users\\ayubi\\SafeRoute\\src\\components\\Navbar\\index.js",[],"C:\\Users\\ayubi\\SafeRoute\\src\\components\\SearchField\\index.js",[],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\ayubi\\SafeRoute\\src\\index.js":"1","C:\\Users\\ayubi\\SafeRoute\\src\\reportWebVitals.js":"2","C:\\Users\\ayubi\\SafeRoute\\src\\App.js":"3","C:\\Users\\ayubi\\SafeRoute\\src\\components\\Navbar\\index.js":"4","C:\\Users\\ayubi\\SafeRoute\\src\\components\\SearchField\\index.js":"5"},{"size":500,"mtime":1608424109636,"results":"6","hashOfConfig":"7"},{"size":362,"mtime":1608424109636,"results":"8","hashOfConfig":"7"},{"size":824,"mtime":1609317987488,"results":"9","hashOfConfig":"7"},{"size":7264,"mtime":1609317827026,"results":"10","hashOfConfig":"7"},{"size":788,"mtime":1609211344172,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"1k0mzzr",{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"C:\\Users\\ayubi\\SafeRoute\\src\\index.js",[],["23","24"],"C:\\Users\\ayubi\\SafeRoute\\src\\reportWebVitals.js",[],"C:\\Users\\ayubi\\SafeRoute\\src\\App.js",[],"C:\\Users\\ayubi\\SafeRoute\\src\\components\\Navbar\\index.js",[],"C:\\Users\\ayubi\\SafeRoute\\src\\components\\SearchField\\index.js",[],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"no-global-assign","no-unsafe-negation"]
