--- src/Pages/SignIn/index.js
@@ -1,19 +1,0 @@
-import React from "react";
-// import Navbar from "../components/Navbar";
-// import SearchForm from "./components/SearchForm";
-// import Navbar from "./components/Navbar";
-// import { BrowserRouter as Router } from "react-router-dom";
-
-const SignIn = function SignIn() {
-
-return(
-    <div>Sign In
-{/* <Router>
-<Navbar />
-<SearchForm />
-</Router> */}
-</div>
-)
-}
-
-export default  SignIn;
